// smsWithMavenForntEnd/src/app/model/student.model.ts
export class StudntAddForm{

   
        sid?: number;
        batchId?: string;
        session?: string;
        stClass?: string;
        stRoll?: string;
        stfirstname?: string;
        stlastname?: string;
        stemail?: string;
        stfathersname?: string;
        stmothersname?: string;
        stpassword?: string;
        strole?: string;
        stdob?: string;
        staddress?: string;
        stgender?: string;
        stphone?: string;
        stPhoto?: string | null;


}